changed so that collision may happen multiple times, but the worm still only dies once

+++ src/player/worm/worm-handler.js
@@ -1,11 +1,15 @@
 module.exports = function WormHandler(eventHandler, shapeModifierI, wormBodyGridHandler, wormBodyImmunityHandler, clone) {
 
     eventHandler.on(eventHandler.events.WORM_MAP_COLLISION, function onWormMapCollision(players, player, worm) {
-        kill(players, player, worm);
+        if (worm.alive) {
+            kill(players, player, worm);
+        }
     });
 
     eventHandler.on(eventHandler.events.WORM_WORM_COLLISION, function onWormWormCollision(players, player, worm) {
-        kill(players, player, worm);
+        if (worm.alive) {
+            kill(players, player, worm);
+        }
     });
 
     function kill(players, player, worm) {
